refactor(event): add basic Event and some refactor
