+++ .changeset/fluffy-spiders-sleep.md
@@ -1,0 +1,5 @@
+---
+"@backend-souls/jscalendar": minor
+---
+
+Initial version with basic Event
