+++ package.json
@@ -1,13 +1,16 @@
 {
   "name": "@backend-souls/jscalendar",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "description": "An opinionated implementation of the JSCalendar defined in the RFC-8984.",
-  "type": "commonjs",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "/dist"
-  ],
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    }
+  },
   "keywords": [
     "JSCalendar",
     "RFC-8984"
@@ -18,15 +21,19 @@
     "node": "20.11.1"
   },
   "scripts": {
-    "test": "jest --forceExit --coverage --verbose",
-    "test:watch": "jest --watch",
-    "build": "tsc -p tsconfig.build.json"
+    "dev": "vitest",
+    "test": "vitest run",
+    "build": "tsup src/index.ts --format cjs,esm --dts",
+    "lint": "tsc",
+    "ci": "pnpm run lint && pnpm run test && pnpm run build",
+    "release": "pnpm run lint && pnpm run test && pnpm run build && changeset publish"
   },
   "devDependencies": {
     "@faker-js/faker": "8.4.1",
-    "@jest/globals": "^29.7.0",
+    "@changesets/cli": "^2.27.1",
+    "@swc/cli": "^0.3.10",
+    "@swc/core": "^1.4.2",
     "@tsconfig/node20": "^20.1.2",
-    "@types/jest": "^29.5.12",
     "@types/node": "^20.11.20",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "eslint": "^8.57.0",
@@ -34,10 +41,10 @@
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-promise": "^6.1.1",
-    "jest": "^29.7.0",
     "prettier": "3.2.5",
-    "ts-jest": "^29.1.2",
-    "typescript": "^5.3.3"
+    "tsup": "^8.0.2",
+    "typescript": "^5.3.3",
+    "vitest": "^1.3.1"
   },
   "dependencies": {
     "date-fns": "^3.3.1"
