+++ src/__test__/product.spec.ts
@@ -1,9 +1,10 @@
 import { randomUUID } from 'node:crypto';
 
-import { describe, it, expect } from '@jest/globals';
+import { describe, it, expect } from 'vitest';
 import { faker } from '@faker-js/faker/locale/en';
+import { Product } from '../attributes/Product';
 
-import { Product } from '..';
+import { classToPlain } from 'class-transformer';
 
 describe('Product', () => {
   it('create a Product with all properties', () => {
