+++ src/attributes/index.ts
@@ -1,0 +1,4 @@
+export * from './Product';
+export * from './Metadata';
+export * from './Details';
+export * from './Link';
