+++ src/domain/Group.ts
@@ -1,0 +1,1 @@
+export type Group = {};
