+++ src/domain/Task.ts
@@ -1,0 +1,1 @@
+export type Task = {};
