+++ src/domain/index.ts
@@ -1,0 +1,3 @@
+export * from './Event';
+export * from './Task';
+export * from './Group';
