--- tsconfig.build.json
@@ -1,4 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "exclude": ["**/*.spec.*", "**/__mocks__/*", "**/__tests__/*"]
-}
